Handle up/down/left/right controls
`event::EventHandler` is nice, but the actual callback function
`key_down_event` doesn't provide the context. I think the intent is
to force implementers to do what I did and control input state with
a struct. In this case, I set `InputState` inputs and let the `update`
function handle how to deal with those inputs, and then resetting them
for next time.

I also changed the `update` function to update the position of the main
sprite based on how much time has passed since the last time update
was called. This means that, should the event loop experience some
lag due to badly structured code, the movement is still a consistent
rate throughout.