Rework all sorts of tile positioning nonsense.
ugh, x/y on rect is *center* of shape. Also 0,0 is bottom left with y
pointing up. I had thought 0,0 was top left with y pointing down, but
that's not the case at all.

New plan for tile rendering involves updating to ggez master where the
new SpriteBatch is available. We should be able to render the track as
one big texture then just translate it. That should be much better than
assembling tiles each tick.